pyproject.toml

[project]

    name = "PyDrawille"
    dynamic = ["version"]
    requires-python = ">= 3.8"
    dependencies = [
        "numpy",
        "pillow >= 10.0.0",
    ]
    authors = [
        { name = "金羿Eilles", email = "EillesWan@outlook.com" },
    ]

    description = "Using Unicode braille characters to draw pixels in console."
    readme      = { file = "README.en.md", content-type = "text/markdown" }
    license     = "Apache-2.0"

    keywords = [
        "console",
        "drawille",
        "unicode",
        "braille",
        "pixel",
        "ascii",
        "image",
    ]
    classifiers = [
        "Environment :: Console",
        "Intended Audience :: Developers",
        # "License :: OSI Approved :: Apache Software License",
        "Operating System :: OS Independent",
        "Topic :: Software Development :: Libraries",
        "Programming Language :: Python",
        "Programming Language :: Python :: 3",

    ]

[project.optional-dependencies]
    dev = [
        "twine",
        "rich",
    ]


[project.urls]

    "Homepage" = "https://github.com/EillesWan/PyDrawille"
    Issues     = "https://github.com/EillesWan/PyDrawille/issues"


[build-system]
    requires      = ["pdm-backend"]
    build-backend = "pdm.backend"


# https://backend.pdm-project.org/build_config/#build-configurations
[tool.pdm.build]
    source-includes = [
        "README.en.md",
        "README.md",
        "LICENSE",
    ]
    excludes = [
        "refant/",
        "./*.otf",
        "./*.ttf",
        "./*.png",
        "./test.py",
        "./test_draw.py",
    ]


[tool.pdm.version]
    source = "file"
    path   = "PyDrawille/__init__.py"


[tool.pyright]
    typeCheckingMode = "standard"
